+++ tests/DependencyInjection/FlagceptionExtensionTest.php
@@ -2,7 +2,9 @@
 
 namespace Flagception\Tests\FlagceptionBundle\DependencyInjection;
 
+use Flagception\Activator\FeatureActivatorInterface;
 use Flagception\Bundle\FlagceptionBundle\DependencyInjection\FlagceptionExtension;
+use Flagception\Decorator\ContextDecoratorInterface;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\Config\FileLocator;
@@ -151,5 +153,32 @@ class FlagceptionExtensionTest extends TestCase
 
         $definition = $this->container->getDefinition('flagception.listener.routing_metadata_subscriber');
         static::assertTrue($definition->hasTag('kernel.event_subscriber'));
+    }
+
+    /**
+     * Test that annotation subscriber is disabled
+     *
+     * @return void
+     */
+    public function testAutConfiguration()
+    {
+        if (method_exists($this->container, 'registerForAutoconfiguration') === false) {
+            $this->markTestSkipped('Only since Symfony 3.3');
+        }
+
+        $config = [];
+
+        $extension = new FlagceptionExtension();
+        $extension->load($config, $this->container);
+
+        $activatorChildDefinition = $this->container->getAutoconfiguredInstanceof()[FeatureActivatorInterface::class];
+        static::assertEquals([
+            'flagception.activator' => [[]]
+        ], $activatorChildDefinition->getTags());
+
+        $contextChildDefinition = $this->container->getAutoconfiguredInstanceof()[ContextDecoratorInterface::class];
+        static::assertEquals([
+            'flagception.context_decorator' => [[]]
+        ], $contextChildDefinition->getTags());
     }
 }